Project 1 - Flappy Bird [Adjustements]
Modified VSYNC value as dt doesn't work properly if you have more than 60 Hz of refresh rate on your screen.

As I have 144 Hz, both gravity and jump worked as if they were "multiplied" making the bird impossible to control.

Added a "FPS Limiter" to the update function in main file to try yo avoid this issue though it ,ight not be the best solution